// Desktop/final-project-yuxuan-weli-main 2/client/src/types/entityTypes.ts
interface AnswerType {
  text: string;
  ans_by: string;
  ans_date_time: Date;
  _id?: string;
  upvoteCount?: number;
  downvoteCount?: number;
}
// ...
interface Tag {
  name: string;
  _id?: string;
}